add section data view

+++ Source/viewLib.h
@@ -9,7 +9,7 @@
 
 #define GetOffByMem(STRUCT,MEM)	((u32)&((STRUCT*)0)->MEM)
 
-void viewNhex(u8* fmt, void* vhex, u8 len);
+void viewNhex(u8* fmt, void* vhex, u32 len);
 void viewNchars(u8* fmt, u8* str, u8 len);
 u32 viewAtoi(u8* numStr);
 u32 viewGetNum(void* numData, u8 islittle);
